Fix the problem that the preview image in Tutorial Tile View is too small
